adding explicit check for log file presence

+++ bin/cycle_logs.rb
@@ -13,12 +13,12 @@ all_logs = Dir.entries("log").keep_if {|l| !l.start_with?('.')}
 4.downto(1) do |i|
   if i == 4
     all_logs.select {|l| l =~ /#{i}/}.each do |log|
-			File.delete("log/#{log}")
+			File.exists?("log/#{log}") ? File.delete("log/#{log}") : next
     end
   else
 		all_logs.select {|l| l =~ /#{i}/}.each do |log|
       basename = log.split('.').first
-			File.rename("log/#{basename}.#{i}.log", "log/#{basename}.#{i + 1}.log")
+			File.exists?("log/#{basename}.#{i}.log") ? File.rename("log/#{basename}.#{i}.log", "log/#{basename}.#{i + 1}.log") : next
 		end
   end
 end
@@ -26,8 +26,10 @@ end
 # find all logs that haven't been rolled over yet and rename
 all_logs.select {|l| l.split('.')[1] == 'log'}.each do |log|
 	basename = log.split('.').first
-	FileUtils.cp("log/#{basename}.log", "log/#{basename}.1.log")
-	File.delete("log/#{basename}.log")
+  if File.exists?("log/#{basename}.log")
+		FileUtils.cp("log/#{basename}.log", "log/#{basename}.1.log")
+		File.delete("log/#{basename}.log")
+  end
 end
